Update LinkHandlerForTMPText.cs
The code checked if the link started with "www" before, now it checks if it's starting with "http://" or "https://" because not all websites start with "www". I had totally forgotten that happens. Thanks to Lucas for pointing that out in the video's comment section!

Plus, "OnClickedOnLinkEvent" is now renamed to "ClickedOnLink", after having had an fun discussion with a friend about picking good names for stuff.